+++ apps/frontend/src/features/admin-form/settings/components/FormSaveDraftToggle.tsx
@@ -30,7 +30,6 @@ const FormSaveDraftToggle = () => {
     <Skeleton isLoaded={!isLoadingSettings && !!settings}>
       <Toggle
         label={t('features.adminForm.settings.general.saveDraft.label')}
-        betaBadge
         description={t(
           'features.adminForm.settings.general.saveDraft.description',
         )}